Close dapui when manually stopping the debugger.
Dapui closes on its own just fine when the debuggee executable
terminates on its own at end of execution but not all debug adapters
reliably send terminate events when debugging is terminated prematurely.
This leaves dapui stuck open with a dead session. Close it ourselves
when hitting one of the "stop debugging" keymaps.