fix: Removed notification plugin.
Foot terminal doesn't seem to like whatever `nvim-notify` is doing to
display a notification popup. A blank window is rendered and remains
stuck on the screen. I need to select the popup window and close it
manually. Notifications are still displayed correctly in the message
buffer as normal while I figure this out.